+++ recognition/HipMRI_UNet_Al_Holliday_43950214/predict.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""
+Prediction script
+
+@author: al
+"""
+
+# TODO: put code here
